really fix average fps

+++ libavcodec/rkmppdec.c
@@ -85,7 +85,7 @@ static float ffrkmpp_compute_framerate(AVCodecContext *avctx)
     static struct timespec reftime;
     static int refframecount;
     struct timespec time;
-    float timediff;
+    double timediff;
 
     RKMPPDecodeContext *rk_context = avctx->priv_data;
 
@@ -96,9 +96,8 @@ static float ffrkmpp_compute_framerate(AVCodecContext *avctx)
         reftime.tv_sec  = time.tv_sec;
         reftime.tv_nsec = time.tv_nsec;
     }
-
-    timediff = ((float)time.tv_sec + ((float)time.tv_nsec / 1000000000.0)) -
-               ((float)reftime.tv_sec + ((float)reftime.tv_nsec / 1000000000.0));
+    timediff = ((double)time.tv_sec + ((double)time.tv_nsec / 1000000000.0)) -
+               ((double)reftime.tv_sec + ((double)reftime.tv_nsec / 1000000000.0));
 
     if (timediff != 0)
         return (float)(rk_context->framecount - refframecount) / timediff;
